buck2: aquery: don't erase target nodes immediately when resolving literals
Summary:
I would like to add more functions in aquery to allow resolving things that
aren't just _the one default output of a thing_.

I don't want to invent a new literal syntax for this and it seems like a bad
idea, so instead I'm going to support things like `all_outputs(cell//foo:bar)`.

To support this, we need to avoid "forgetting" about the literals and instead
we want to resolve them to multiple things: the current action nodes we have,
and a special node that represents the target.

This does mean that if you do a aquery you'll now get both, but that's not so
different from what you would get when doing `cquery` where you might get a
`forward` node for the thing you are querying.
 ---

Note that this requires doing some annoying things with the "NodeRef". Ideally
this would return an enum and not a `&Thing`, but refactoring that (I tried)
is very painful and doesn't seem like an overall improvement in the codebase,
you end up with trait definitions looking like ... this:

```
pub trait LabeledNode: Dupe + Send + Sync + 'static {
    type NodeLabel: for<'a> NodeLabel<Ref<'a> = Self::NodeRef<'a>>;
    type NodeRef<'a>: NodeRef<Self::NodeLabel> + 'a = &'a Self::NodeLabel;

    fn node_ref<'a>(&'a self) -> Self::NodeRef<'a>;
}
```

Reviewed By: stepancheg

Differential Revision: D47958345

fbshipit-source-id: 4461503d490357680f73e67778520ecf9e923bbb