buck2: chrome trace: change the bucket duration to 100ms for counters
Summary:
Because honestly you're not going to be able to tell the difference when
looking at the counters in a UI (and if this really mattered, we could make
this configurable).

This sort of requires the previous diff (that was part of the motivation),
since for the bucketing to impact the volume of data we produce, it first
needed to apply to everything.

Reviewed By: lmvasquezg

Differential Revision: D48025945

fbshipit-source-id: 78dd1c4a17f24bf552e40f006f3f7dece9c720ac