buck2: chrome trace: make the rendering of zeroed counters more efficient
Summary:
When we emit a Chrome trace, we try to not send values for counters that are
currently zero. To make this work and avoid the renderer interpolating when we
go from zero-but-omitted to non-zero, we need to emit an extra event that sets
the value to zero just before.

However, right now we do this whenever a counter goes from zero to non-zero,
but we disregard the fact that we normally buffer the counters we emit. So, if
during once buffering window, X goes 0 -> 1 -> 0 -> 2, we'll emit the zeroes
twice, which is kinda redundant.

Reviewed By: lmvasquezg

Differential Revision: D48025944

fbshipit-source-id: 3fa7a2fc75f9da8fd9ad76003d8a341a073998d5