Add an `_apple_platforms` dictionary to all rules
Summary:
With the current setup of target platforms and configuration, we have no way to _discover_ target platforms and know which to use in which environment. We're only able to change buck config and get the default for interested targets. More [info here](https://fb.workplace.com/groups/buck2dev/posts/3428993270722054).

While we've worked around this so far, this makes it extremely difficult to get the tests and test deps for a non-default configuration (arm64) without knowing those tests in advance.

The long term proposed solution to this problem is to use [configuration modifiers](https://fb.workplace.com/groups/buck2dev/posts/3466287650325949). However, until then, we'll need to a temporary way to get the target platforms for a particular app in a specific configuration.

Our target platforms for iOS targets are either:
- the default from the mappings (https://www.internalfb.com/code/fbsource/[617d47800ec4dac2960ac3d447b98200cfc18f16]/tools/build_defs/target_platform_defs.bzl?lines=97-126)
- or generated off the [`config_backed_target_platform`](https://www.internalfb.com/code/fbsource/[617d47800ec4]/tools/build_defs/apple/target_platforms.bzl?lines=60)

Thus, whenever we set the target platform for an iOS target, we can also set a dictionary of the possible options, which are queryable.

Due to how target platforms are used in the macro layer, adding this attribute to all rules. We're simply setting a defined platform dep, which after measuring memory consumption for a large graph (ex: fbios) appears minimal.

Ex:
```
buck2 uquery fbsource//fbobjc/Apps/LightSpeed:LightSpeedApp -a default_target_platform -a _apple_platforms
{
  "fbsource//fbobjc/Apps/LightSpeed:LightSpeedApp": {
    "default_target_platform": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-x86_64-local",
    "_apple_platforms": {
      "iphoneos-arm64-local": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphoneos-arm64-local",
      "iphoneos-arm64-master": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphoneos-arm64-master",
      "iphoneos-arm64-production": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphoneos-arm64-production",
      "iphoneos-arm64-profile": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphoneos-arm64-profile",
      "iphoneos-arm64-rc": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphoneos-arm64-rc",
      "iphonesimulator-arm64-local": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-arm64-local",
      "iphonesimulator-arm64-master": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-arm64-master",
      "iphonesimulator-arm64-production": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-arm64-production",
      "iphonesimulator-arm64-profile": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-arm64-profile",
      "iphonesimulator-arm64-rc": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-arm64-rc",
      "iphonesimulator-x86_64-local": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-x86_64-local",
      "iphonesimulator-x86_64-master": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-x86_64-master",
      "iphonesimulator-x86_64-production": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-x86_64-production",
      "iphonesimulator-x86_64-profile": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-x86_64-profile",
      "iphonesimulator-x86_64-rc": "fbsource//tools/build_defs/config/platform/apple:LightSpeedAppBase-iphonesimulator-x86_64-rc"
    }
  }
}
```

Reviewed By: rmaz

Differential Revision: D47920469

fbshipit-source-id: 9464121663f08f22251483d2a4e6b48f0d102a6e