Use client builder for top-level client configuration
Summary:
This diff replaces the `http_client()` toplevel call that's used to configure the client in daemon state init and in the manifold wrapper client with one driven by the new client builder.`

While doing this, let's take advantage of passing in the root config so we can drive configuration of different aspects of the client (like new timeout parameters).

Sadly, still need to pass the `allow_vpnless` flag because that comes from the daemon init config.

Reviewed By: krallin

Differential Revision: D47955388

fbshipit-source-id: 88f90666210b3d165fd29b3ccc90d8107510a9d4