Refactor http client construction
Summary:
This is a fairly significant refactor of how we construct http clients for buck. The key distinction is that we no longer have wrapper types that all impl the http client trait. With this diff I'm posing that we don't actually need that (IMO complicated) design.

## A motivating example
This complicated wrap-another-client approach has led to at least on bug with the following sequence, the net result being vpnless builds relying on URLs that will never work off VPN taking forever (because they have a long timeout, probably from x2pagent?):
1. Buck starts with `allow_vpnless=true` on a laptop gated into vpnless.
2. We use the [vpnless unix proxy client](https://www.internalfb.com/code/fbsource/[67d603144a76da37dccf7c5356fefd2f4f251fef]/fbcode/buck2/app/buck2_common/src/http/x2p.rs?lines=97)
3. Buck does an http action (`ctx.actions.download_file`) and hits [`X2PAgentUnixSocketClient::request`](https://www.internalfb.com/code/fbsource/[67d603144a76da37dccf7c5356fefd2f4f251fef]/fbcode/buck2/app/buck2_common/src/http/x2p.rs?lines=123) which calls [`X2PAgentUnixSocketClient::request_impl`](https://www.internalfb.com/code/fbsource/[67d603144a76da37dccf7c5356fefd2f4f251fef]/fbcode/buck2/app/buck2_common/src/http/x2p.rs?lines=112).
4. This correctly mutates the URL scheme to be `http` (which is mandated by x2pagent).
5. When fetching a URL that doesn't work off-vpn, say, `"https://our.intern.facebook.com/intern/managed_sdk/component/921533932454562/"`, we get a 301 redirect to the generic FB login page (because again, on vpnless we can't fetch this page).
6. The [redirect processing logic in the base client runs](https://www.internalfb.com/code/fbsource/[67d603144a76da37dccf7c5356fefd2f4f251fef]/fbcode/buck2/app/buck2_common/src/http/secure_client.rs?lines=88-92), but crucially, _doesn't_ rewrite the scheme. This results in... a long processing time? from x2pagent.

What _should_ happen instead is that we download the redirect page html, try to hash that, and get a hash mismatch. The error message in this case sucks (will be improved later in this stack), but regardless having a build sit there spinning its wheels due to (what is probably a bug in x2pagent) isn't great. I'll report this bug to the x2p folks so they at least fix it on their end.

This architecture of wrapping different clients got us here, and it's difficult to trace logic through the different wrappers. It's also difficult to extend if you e.g. want to apply different bits of logic along the request path. An example would be introducing a layer to handle connect and read timeouts; to do this today requires updating all the different client wrappers we support. This feels... not great?

## What do we do instead?
Instead, we have basically three key kinds of clients we want to support:
* Regular OSS client that optionally sets standard environment proxy vars if set
* Internal http client that uses Meta TLS certs for auth
* x2p clients that can either route through an http proxy OR through a unix domain socket.

The key difference between all of these is not necessarily the client, but the _connection_. So let's make it possible to construct different connections but re-use the same client logic for all of them.

The only place this new design breaks down is for x2p clients where we have to override the uri scheme. This is a small enough change that this is baked into the client request flow and done if the client is configured for vpnless (which itself is a separate attribute on the client).

Reviewed By: krallin

Differential Revision: D47955386

fbshipit-source-id: 85f4e0ccb1ba37ad6c4244c0b34962a203ed3e2c