Acquire worker permit earlier, without moving initialization
Summary:
Fixes the same 2 issues as D47709827 (waiting for a permit during worker execution prevents actions from being cancelled before they start, and waiting counts towards action execution time) without causing worker initialization to happen before inputs are materialized  (which breaks builds unless the worker itself happens to already be materialized).

Experimenting with setting weights instead of a worker concurrency limit (D47753011) shows an improvement in cases where the effective limit is similar (because it avoids the bug with cancellation), and in much smaller machines (macbooks), but it shows a significant slowdown in other cases - specifically large builds on 32 core devservers.

If we end of setting a lower weight than 5, we may want to also set a fixed cap to optimize perf on large machines since those are more likely to overload the worker, not the machine. This requires some more experimentation, but probably makes sense to fix this anyway.

Reviewed By: krallin

Differential Revision: D47935028

fbshipit-source-id: ab9685d7ae04ca3478f950b4f7d134cc2be57c88