Consider profiling when building artifacts
Summary:
## This stack
See D47987891.

## This diff

Change the names of generated artifacts (e.g. package directories, lib names) to consider the status of profiling.

This is needed to (a) make it clear for every artifact if it was build with or without profiling and (b) to fix buck2 errors from declaring outputs with the same name of running the actions with the same category name (e.g. "haskell_link_...").

Reviewed By: helfper

Differential Revision: D47988498

fbshipit-source-id: 16a8e7da2e3009b76623fe2c2bc30ff091aa9665