buck2: allow clients to specify extra metadata to inject into logging
Summary:
Today they already can do this via `--config scuba.defaults="{\"normals\":
{\"foo\": \"bar\"}}"`, which is a rather shitty way to do it because:

- It invalidates configs
- It can be literally anything
- It's hard to grep for

This diff provides a more opinionated way to do this. You set
`--client-metadata=foo=bar`, and then that will get sent straight to our
logging as `client_metadata_foo=bar`.

Concretely, the reason I want this is to allow MyBuild to inject something into
buck2_builds that tells us "this is the benchmarked build, and not the warm up
build", but this might have applications for others (e.g. as a way to replace
`-c client.id`).

Reviewed By: IanChilds

Differential Revision: D48033517

fbshipit-source-id: e3b62071e3e4d41437ed3b1d08ef7b68cb68d826