buck2: check for stale cwd when you start
Summary:
Capture the cwd when we start, then when we run a command, check if the inodes
match.

If they don't, then buck-out has probably gone away (or might have been
remounted), and we need to restart the daemon (changing our working dir isn't
quite enough, since we hold some stateful handles to notably the materializer
state).

See here for an example of where this is relevant:

https://fb.workplace.com/groups/buck2users/permalink/3474748969448126/

For now I'm just making this a soft error but my plan is to turn this into
one we auto-restart from.

Reviewed By: stepancheg

Differential Revision: D48065551

fbshipit-source-id: b6c0b75c47fc973a8dc0e6e4e847273ffbfd900c