Modern dice to have variants in its computation ctx
Summary:
Introduce an enum for the ComputationCtx of ModernDice.

This is used to support the `compute_many` api in D47936925

With compute many, we have a set of lambdas that each receive a `&mut ComputationCtx` to create a future. For the parallel-ly computed futures, each of the underying `ComputationCtx` will hold a different data that the standard data that a key receives as its `ComputationCtx`. So, we introduce this enum to allow us to distinguish the regular computation ctx for keys, vs the context for each parallel computed lambda

Reviewed By: cjhopman

Differential Revision: D47983561

fbshipit-source-id: ffda3e6897083a34e3fd146344e7298d02e60770