Move Android min_sdk_version constraints to prelude
Summary:
The Android min_sdk_version constraints currently live in xplat, but there is a designated directory
within the buck2 prelude for android constraints.

There's not a clear reason for these values to live in xplat, but it does introduce a cross-cell
dependency from the buck2 prelude, so these constraints seem like they might be better suited for
the prelude.

Reviewed By: IanChilds

Differential Revision: D48122467

fbshipit-source-id: 1a4e2739a761dc510df17e9a9fe9187057569bf5