Plugin deps: Add `attr.plugin_dep()`
Summary:
First step of adding plugin deps. Adds `attr.plugin_dep()` and plumbs it through everything, as described in the RFC. For now we just completely ignore these at configuration time, doing something with them is the next step.

Unfortunately, we can't make `PluginDep` be a `DepAttrTransition` because that type is used for configured deps as well. This means we need a new coerced attr type which adds a bit of boilerplate.

Reviewed By: scottcao

Differential Revision: D47859962

fbshipit-source-id: ec420ddb55f488daa1d260556319fde3f84d302b