Add a new Owned and Ref type representing un-lifetimed reference relationship
Summary:
This is used to support the compute_many api in D47936925

With compute many, we have a set of lambdas that each receive a &mut ComputationCtx to create a future. For the parallel-ly computed futures, each of the underying ComputationCtx will hold a different data that the standard data that a key receives as its ComputationCtx.
The parallel ctx will live shorter than the base context, and it will refer to data held by the base context. However, we don't want to introduce a lifetime on the references, as it would result in adding extra lifetime for `DiceComputation` and propagating an extra unimportant lifetime everywhere and making api verbose.

The following proposes a Owned and Ref semantic that doesn't have lifetimes (so slightly unsafe). However, it mostly uses Arc and Weak ptr underneath to maintain some safety while limiting the API to be a owned/ref relationship. Most access on the Ref results in Option for some safety

Reviewed By: cjhopman

Differential Revision: D47983560

fbshipit-source-id: 52e442ae4387375e8579f50a822fce931439559a