buck2: also create BUCK_SCRATCH_PATH on RE
Summary:
Like it says in the title. Make this an input and ensure this gets created.

Note that we are now passing the scratch path in 2 different ways to the local
executor.

This is because I want to keep the action hashes unchanged unless the feature
is enabled, but I need to not break the existing code that wants TMPDIR, so I
need the path to be passed even when the feature is not enabled.

The rollout will be:

- Land this stack up to this diff
- Turn the feature on (invalidate caches)
- Land the rest of the stack after this diff

Reviewed By: wendy728

Differential Revision: D48120895

fbshipit-source-id: a2b048d4a69d31eb4c6a0a7b74c488692fd34b8d