Disable starlark stack tests when building with asan
Summary:
Attempt number 2 to address the issues with the buck2 tests that look at stack frame size and growth direction. With D47566299, which re-enables sanitizers when building with buck2, I found that the stack tests were failing. Sprinking more `std::blackbox::hint` did not help, no matter what I tried.

So failing that, I fell back to using inline assembly to get the stack pointer, which appears to work on all platforms that support `asm!()`. Open to any alternatives here, the tests are easy to reproduce.

Reviewed By: stepancheg

Differential Revision: D47571173

fbshipit-source-id: 93830af2656b3591eccc4956b8c237d30034c862