Make `sub_target` and `configured_sub_target` available outside of bxl
Summary:
My goal with this diff is two-fold:

 1. Make these functions more discoverable by putting them onto the type, instead of on globals
 2. Make them available outside of bxl

For the first one I feel reasonably strongly that this is an improvement. Especially because these names are not all that obvious imo.

The second one I feel less strongly about. I also don't have an immediate use case. Still, someone will probably want to extract some logic out of their bxl into a bzl and then run into this sooner or later?

## Naming

After some discussion in the core sync, the agreement was to use "subtarget" oriented names, not "providers." I chose to use `with_sub_target` over `sub_target` as I think that more clearly communicates that this is a type change

Reviewed By: krallin

Differential Revision: D47941237

fbshipit-source-id: 19ba7ec08939418167e5ffc82b44ba039ab984d6