Make file_expr async and use of DiceComputations explicit
Summary:
We want to make all uses of DiceComputations &mut

previously bxl had all uses of DiceComputations hidden and sync.
The goal is to make these less implicit so that there's a few entry points at the top level to get the `mut DiceComputation` instead of having anywhere access it as we do now.

Reviewed By: JakobDegen

Differential Revision: D48092482

fbshipit-source-id: 565a3ee877da3952f05c057abca292307e7a5504