Add --show-simple-output build flags to Buck2
Summary:
Adds two new flags
- `--show-simple-output`
- `--show-full-simple-output`

These will print the outputs one per line without prefixing them with the target.

A very common thing to do prior to this change is
```lang=bash
./run_script --file="$(buck2 build //path/to:target --show-output | cut -d' ' -f1)"
```

This reduces the complexity of these scripts and removes the need for potentially platform specific tools to extract the outputs. This is particularly useful when building just one target or if you just want an unordered list of outputs from all targets.
```lang=bash
./run_script --file="$(buck2 build //path/to:target --show-simple-output)"
```

Reviewed By: krallin

Differential Revision: D48165534

fbshipit-source-id: 0ebc0a1485910bf31936c89842b78e9a8e840ce3