Centralize logic in bxl
Summary:
I don't think it's helpful to be going back and forth between bxl and Rust. We can just generate all the things we need in bxl at once. On the other hand, going back and forth leads to duplicate code and is a refactoring hazard because of the out of sync MSDK bumps.

Also, this makes the rest of the stack easier.

Reviewed By: davidbarsky

Differential Revision: D48096437

fbshipit-source-id: 70278868650f7ae11675d5d0eec4c6591d0ecc84