Remove some 'v lifetimes in artifacts.rs
Summary:
this `'v` is incorrect as it forces the dice computation to live as long as the artifact group, and also implies that the future returned from dicecomputation lives the same duration.
This is not true, as we actually separately dupe items for the dicecomputation future.
So, kill the lifetime bound and use the anonymous inferred lifetimes, which are more correct

Reviewed By: cjhopman

Differential Revision: D48279520

fbshipit-source-id: 36c5e0b698c42747563eb6638b896e7be25372cb