Exclude APK-under-test resource packages from test APK
Summary:
This is the `buck2` counterpart to D20853045.

We keep track of all the `package` attributes that are specified on `android_resource` rules, and we don't allow any of those to go into the test APK, since they can override the classes from the APK-under-test and cause fields to not be found.

Reviewed By: christolliday

Differential Revision: D48316795

fbshipit-source-id: a5737508cd9606090d3de801e25c44ffd126589c