Add `_workspaces` attribute
Summary:
## Goal of this stack

The projects generated by rust-project currently have the following two related problems:

 1. They never include any rdeps of the target being edited. This is unfortunate, because if I hit "find all references" on something in `buck2_interpreter`, I want to also find references in `buck2_interpreter_for_build`.
 2. All targets except for the one being edited are marked as being "not in the workspace." This means that r-a will not automatically pick up changes to them, which is quite annoying.

This stack proposes to fix both of these problems by introducing a concept of workspaces. Imagine a user with a `myproject/` directory, containing `lib1`, `lib2`, and `bin`. They'll be able to create a `myproject/PACKAGE` file with:

```
load("prelude//rust:add_workspace.bzl", "add_rust_workspace")
add_rust_workspace("//myproject:bin")
```

Whenever the user edits a target that is a) within that package, and b) a dep of `:bin`, the generated rust-project will be modified to:

 1. Include `:bin` and all of its deps, even if `:bin` was not the target being edited.
 2. Mark all deps of `:bin` that are within the package as being within the workspace

The goal is that this fixes the issues mentioned above.

### This diff

This diff implements a way of getting the workspace information onto the targets. For now, we use a special attribute that stores a list of labels. What we actually want is metadata (see D47129414), but that hasn't landed yet. I'll convert it when metadata lands.

Reviewed By: davidbarsky

Differential Revision: D48096435

fbshipit-source-id: 5871c3fc35675d291f2b650ca4f75cc15a9b0d01