Allow ensured artifacts in user event log
Summary:
It would be helpful to be able to write the ensured artifact path to the user event log, which end users can use to determine where certain files are.

This implementation does something similar to `ctx.output.print()` - `instant_event()` will parse ensured artifacts using a helper method to get the abs or rel path.

Reviewed By: themarwhal

Differential Revision: D48315026

fbshipit-source-id: d78031c0066a8a939c302b06b1c5f546f02c89b8