Redo signatuer of via_dice more for rust lifetime inference
Summary:
Because rust doesn't yet have a `for<'a> where 'a: ...`, in order for rust to infer the correct lifetime bounds on the higher rank bound, we need the arguments the closure uses to express bounds, which we do so by adding lifetime parameterizations to the type that the closure accepts. Specifically, by changing the closure for accepting a `for <'x> &'x mut DiceComputation` to `for <'x> &mut DiceComputationRef<'a>`, rust will infer a `'a: 'x` bound on the `for <'x`>.

The reason we introduce a new type instead is because we don't want to add lifetimes to all DiceComputation objects as it's generally not useful, so, just add it to the parallel instances by creating a new `DiceComputationRef<'a>`.
This derefs into a normal DiceComputation so it can be used just like one.

Reviewed By: cjhopman

Differential Revision: D48254304

fbshipit-source-id: 080f3f3d459bbe0aab4dc666d6ae4ca8615463c3