Fix bug in verifying execution platform within dynamic output
Summary: We were reassigning `target_platform` in line 919 on the old file, which would return something if the named parameter was `None` and the global target platform was `Some`. Thus, when we validate the execution platform resolution for dynamic output `bxl_actions()`, the target platform var would be `Some`, and this would trigger the error saying that the execution platform needs to be inherited from the parent.

Reviewed By: zertosh

Differential Revision: D48432476

fbshipit-source-id: 49fd5088cdd6932e1030fcf5e3ccb97e20bbdb13