Add basic support for per-source auto-PCH
Summary:
This adds opt-in support to automatically generate a PCH for each C++
source (based on bmaurer's investigation in this area) -- by extracting
its leading `#include`s into a header and pre-compiling it -- which can
dramatically improve the compile-cycle time when iterating on `.cpp`s
~50% (the subsequent diff in this stack uses `-fpch-codegen`/
`-fpch-debuginfo`, which makes this faster w/ tradeoffs, as it can eagerly
generate more code than w/o PCH).

In basic tests, this improves the `.cpp` iteration workflow ~50% while
regressing the `.h` iteration workflows by up to ~30% (presumably due to
the overhead in serializing/deserializing the AST).  A followup here is to
support some way to "race" both approaches in incremental builds to get
the best of both worlds.

TODOs:
1) We currently re-generate the source to strip the leading `#includes`
   that were put in the PCH, as using the original source fails
   compilation w/ conflicts.  We need to figure out if theres some way
   to avoid this or, if it is needed, make sure we're reporting the original
   `.cpp` in error messages and not breaking cases of source-relative
   includes.
2) The `#include` extractor utility is really basic and likely needs some
   work (e.g. it can't walk past `#if*`s).

Differential Revision: D46329995

fbshipit-source-id: e9d1a83ab0162e53da8adaa15d4a74b38e865c26