Remove 'c on NodeCache
Summary:
This captured `'c` which is the lifetime of the DiceComputation.
Once we make DiceComputation `mut`, then capturing this lifetime would prevent the cache from being sharable across different parallel invocations of query, as the lifetime of the mut DiceComputations are not the same.

So, we must remove this lifetime.

Reviewed By: JakobDegen

Differential Revision: D48279522

fbshipit-source-id: 6fe7590c65a37ae499afdc92c80dfcdd400fcce1