dep_file: Check for an identical action match in the local dep file cache before action cache
Summary:
Discussed in internal chat (Buck2 dev).

A few months back, we swapped the order of local dep file cache lookup and action cache lookup. This was because the local dep file cache miss can be expensive as we have to materialize and inspect possibly very large dep files, while action cache lookups are relatively cheap.

However, when we hit the case where the action stored in the dep file cache is identical to the one at hand, we actually don't have to materialize the dep file, making it a very cheap cache hit. This case is better than the action cache hit case.

So the diff changes it from the following
```
1. action cache lookup
2. local dep file cache lookup (full comparison with depfile inspection)
3. fall back to execution
```
to the following
```
1. local dep file cache lookup (partial for the identical action case)
2. action cache lookup
3. local dep file cache lookup (full comparison with depfile inspection)
4. fall back to execution
```

Reviewed By: cjhopman

Differential Revision: D48432813

fbshipit-source-id: e183ffc72fee7f6b16b2222f16b0235984b555cf