Provide ability to set `stripped` on target which overrides global default
Summary:
**Problem**

At present, `apple.stripped_override` overrides the stripping for *all* targets, regardless if they have an explicit `stripped` attribute. This is problematic because it prevents the ability for targets to opt-out out of this behaviour.

**Solution**

We make `stripped` behave as expected - if left unspecified, the default value is picked from `apple.stripped_default`. If `stripped` value is present, then it's the source of the truth.

NOTE: We're renaming `apple.stripped_override` to `apple.stripped_default` to better reflect the semantics.

Reviewed By: blackm00n

Differential Revision: D48555493

fbshipit-source-id: f9f3b639d4c2a064a980c7a8cb61c2ef20dd9af2