Move compiler typechecker to the end of bzl file
Summary:
Invoke typechecker in the end of evaluation of bzl file.

Instead of evaluating after "last statement referencing types".

I figured out how we should typecheck.
* for each top-level statement
  * typecheck this statement alone but without inner content (i.e. typecheck `a: T = b()`, but do not typecheck inners of `def`)
  * evaluate the statement (either properly in compiler or lightly in lint)
* then for each top-level def
  * typecheck it having globals evaluated/typechecked in the previous steps

With previous approach, compiler typechecker did not see the problem in this code for example:

```
def foo() -> int:
  return X

X = "a"
```

This diff is a step into that direction.

The diff is large, but basically it moves

https://www.internalfb.com/code/fbsource/[a994dd5cb82432029ae6c865faa5a71cd4979acf]/fbcode/buck2/starlark-rust/starlark/src/eval/compiler/module.rs?lines=168-170

to the end of the function, and removes dead code.

# Also

Also it greatly simplifies code (and semantics of typechecking).

# Also

I needed it for typechecking of records.

Explanation is hard, ignore if you don't get it. Previous reasons are enough.

"Solver" tries to infer types of all variables. Including types of globals. Compiler typechecker evaluated:

```
MyRec = record()
```

So compiler typechecker knows precise type of `MyRec`. But solver "solves" it, and overwrites it with generic function.

Following diff D48459899 changes typechecker so that when solver works, all globals are already resolved and types for them are not inferred.

Reviewed By: iguridi

Differential Revision: D48459901

fbshipit-source-id: bc382c6a6e039bf64bdf75cac287bb7fd375c570