fix outdated link in buck2 writing rules documentation
Summary:
In the "workflow by example" section, the first listed link points to a file containing only a comment saying all attributes have moved to `prelude/decls`. This update changes the link to follow that comment exactly, although it's unclear to me if `prelude/decls/common.bzl` is a better destination.

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: ndmitchell, shonaganuma

Differential Revision: D48569255

fbshipit-source-id: 6921392bf737ea317c4ca785d7b34ddee7222c8e