No longer rely on the absolute path prefix
Summary:
We have APIs available to get the absolute path, and in prior diffs, have updated the rules to use them.

This allows us to stop reading this buckconfig and cleanup the API. We can now also always define the `abs-argsfile` target, and use `use_absolute_paths` in argsfile generation to state which type of argsfiles we're defining.

Reviewed By: blackm00n

Differential Revision: D48448858

fbshipit-source-id: 3f7c59a177addfbe28c0a4fa302f7b63970d1159