remote_dep_file: Modify CommandExecutionRequest/Response for remote dep file lookup
Summary:
Change the `CommandExecutionRequest` and `CommandExecutionResponse` to tack on some extra data.

For the request, we want to include the remote dep file key as that will be used as the key into the remote dep file cache look up. If the value is None (no dep file for the action), the cache lookup will be skipped.

For the response, we propagate back the DepFileEntry proto value that is stored in the cache as an action auxilary data, which was added back in D47027571 and D47402076. This will be used to perform the dep file validation step, which involves looking at the contents of the depfiles and ensuring the important tagged inputs have not changed.

Reviewed By: scottcao

Differential Revision: D48595640

fbshipit-source-id: cc8e4d6b570e8fd630f590b764d31d25a7254cc3