Increase DFA size limit for filter function
Summary:
Default DFA limit 2M

* `dfa_size_limit` sets hybrid cache capacity? ([source](https://github.com/rust-lang/regex/blob/81e328a29f8c57cb3622930104177e8606270230/src/builders.rs#L189-L190); does not make sense)
* [default value 2M](https://github.com/rust-lang/regex/blob/81e328a29f8c57cb3622930104177e8606270230/regex-automata/src/meta/regex.rs#L3093-L3094)

so default 2M limit seems to be not enough, evaluation switches to NFA (saw it in profiler), and NFA takes forever. Increase the limit to 100M.

Only for `filter` operation for now to unblock iOS migration.

Something is very wrong with `regex` library, we have third issue with it in the last couple weeks.

Anyway, I think we may:
* ditch "fancy" regex (prohibit backtracking)
* fail if regex cannot be evaluated as DFA with given memory limit

Anyway, not sure what should be long term fix, for now I just want to unblock iOS migration.

Reviewed By: cjhopman

Differential Revision: D48633746

fbshipit-source-id: 8c427b67bf87ec53e6f0135fefa976859d5d65e4