TyBasic::Callable
Summary:
Separate callable from `TyCustomFunction`.

The idea is this: every function is "callable". Some non-functions are callable. But concrete function types do not intersect even if they have exactly the same signatures.

For example, type of `min` and type of `max` functions do not intersect.

Practically, this is to make record types (which are callable) do not intersect.

Currently callable is parameterless, because it can only be created with `typing.Callable`. Can be extended later.

Reviewed By: wendy728

Differential Revision: D48522581

fbshipit-source-id: 1bf47174f243d4948b7e1ff8a7887c7c7bf36085