starlark_syntax crate
Summary:
`starlark` crate is annoyingly slow to compile.

This stack tries to pull parser from the crate into a separate crate. One of the offenders is `grammar.lalrpop` which expands to a 20K LOC `.rs` file (P812222556).

Top of the stack at the moment of writing (D48578341) pulls lexer out of the crate.

D48724524 patches ShipIt.

Reviewed By: JakobDegen

Differential Revision: D48568733

fbshipit-source-id: aaa83ec2da42c190551522ca3fd4924ffcb75ba2