Evaluate globals in lint typechecker; do not infer types
Summary:
We have lightweight evaluator of the module (since D48297380) for the lint typechecker.

Previously it only populated type expressions.

Now it also evaluate globals (values or types).

"Solver" now uses evaluated globals instead of inferring their types. Global are supplied from:
* real globals after bzl evaluation in compiler typechecker (possible since D48459901)
* partial globals from lightweight evaluation

Lightweight evaluator is dumb (it was implemented for type evaluation). So it no longer sees the problem in this code now:

```
def foo() -> int:
  return 1

X = foo()

def bar() -> str:
  return X
```

I think it is OK to keep it this way for now.
* if it is not, we can support more expressions in lightweight evaluator
* or postpone it until we make typechecker staged as described in D48459901

Reviewed By: ianlevesque

Differential Revision: D48459899

fbshipit-source-id: f43701e863b52ceaeaf00f47d70a2ae05301ea98