Be able to register artifacts with ctx.actions.anon_targets
Summary:
As the title says. Basically we are porting over the promise artifact registry logic over to anon target registry. End goal is to have promise artifacts be resolved with the analysis ran for the anon target, and not for the caller/consumer's analysis registry.

*Note, this does not handle short path assertions yet.* You can call them, but they are not asserted to be correct anywhere.

This behavior will be controlled by a temporary `with_artifacts` flag (see next diff) to help with the migration. I will migrate all uses throughout the repo after all this is landed.

Reviewed By: cjhopman

Differential Revision: D48568065

fbshipit-source-id: 18587069930258fa12e33a4176f7b01f809324e7