Don't pass current directory to parse import if relative path is disabled
Summary: Currently `parse_import_with_config` requires passing in  `current_dir` even when relative import is disabled with `allow_relative_import=False`. This diff refactors `allow_relative_import` argument to be an enum instead of a boolean so that `current_dir` only needs to be passed in if relative import is allowed. This allows me to call `parse_import_with_config` in the following diff without passing in a `current_dir`.

Reviewed By: JakobDegen

Differential Revision: D48716225

fbshipit-source-id: cedab5d715f5559c03bf0f6c98987fd52c0e38c4