buck2: test: guard against cycles in configuring labels
Summary:
We'll need this in the next diff. Add it. Note that we separate being asked for
a label with skip and without skip. This ensures that if we're asked for a
label with skip then the same label without skip, we report the incompatibility,
at the expense of perhaps a bit of extra unnecessary work.

Reviewed By: iguridi

Differential Revision: D48683127

fbshipit-source-id: 59bac2182462b6575b0071ed1ede9d7c59bd6f8b