fix: jump to definition on dotted definition
Summary:
In cases where you try to go-to-definition on a dotted identifier, e.g. `foo.bar.baz`, the LSP will try to find where `foo` is defined, and check if it's a `struct` definition with a property named `bar`, and jump to that. However, in cases where this location cannot be resolved, an `unwrap_or_default` would result in jumping to to the top of the file. Instead, in such cases, jump to the definition of the variable `foo` itself, which is better than jumping to the top of the file.

Fixes https://github.com/facebookexperimental/starlark-rust/issues/90

X-link: https://github.com/facebookexperimental/starlark-rust/pull/92

Reviewed By: JakobDegen

Differential Revision: D48863997

Pulled By: ndmitchell

fbshipit-source-id: 07015bd7484d0c2d5e74544a9a91fcfbbd76c046