buck2: materializer: deferred: buffer access time updates
Summary:
From krallin original version:

I see this in profiles, and it bothers me a bit. This doesn't really have
much influence on the runtime of anything, but those are trivial to batch
and probably should be.

Generally, we try to have the work the materializer in SQLite does be
proportional to the I/O it does, and this one isn't, because it might
correspond to zero I/O, so let's batch it to reduce its cost.

Reviewed By: krallin

Differential Revision: D46561039

fbshipit-source-id: fa6ca3ea28e5c3d84a1488dc9ca968e6864fad23