Make TyUser.matcher optional
Summary:
Some types cannot appear in type expressions. For example, for

```
Foo = record()
```

type of `Foo` cannot appear in type expressions (type of `Foo` is unnamed). Symbol `Foo` can appear in type expressions, but that is type of instance of `Foo`, not type of `Foo`.

Preparation to convert `TyRecordType` and `TyEnumType` to `TyUser`.

Reviewed By: ianlevesque

Differential Revision: D48857360

fbshipit-source-id: e9170fff694c4579ef24ce64482da70f06b84f1f