Remove `error_messages` field from `CqueryResponse`
Summary:
Cquery errors are normally reported by the `error_messages` on `CommandEnd`. There is a field on `CqueryResponse` though that was also used for error messages - the only errors that would be reported here were errors that occurred trying to print the result to stdout.

This is bad for two reasons: 1) It's confusing. 2) These errors never made it to ingress or into our telemetry

Reviewed By: iguridi

Differential Revision: D48899942

fbshipit-source-id: 873449d0755adc981878670620de0cdeeb9c92f2