Implement access times setting
Summary: If access times are diabled, then let's not keep touched artifacts in a buffer. But let's also warn the user of this when trying to flush all access times using the dferred materializer extension. Later we can changed this to have multiple options inc ase we want to flush but only when a command is not running. For now let's just use a bool.

Reviewed By: krallin

Differential Revision: D48779173

fbshipit-source-id: a0a14ec59400d28f6f17b8489bc3eaf5fdb025a7