Refactor Julia rules
Summary:
This diff (finally) addresses a suggestion made by krallin. Specifically, we refactor the `julia_binary` rule such that it stores all important information in the json file it sends to the python parser used to build the final julia command.

Previously, we were relying on both the json file (for the library information) *and* argparse to communicate relevant details. This resulted in a rather cumbersome implementation.

The refactored version is much cleaner.

Reviewed By: krallin

Differential Revision: D47134081

fbshipit-source-id: 7219d79cbfeb1d3a468ee20460f8e72ce58e87b1