New generic request: simpler client-server interop
Summary:
I love protobuf when backwards compatibility is needed, but for the most commands, it is not needed.

Let's pass requests and responses as Rust struct serialized with serde using universal one gRPC method.

Most commands are the same:
* single request
* streaming response
* client context
* no partial result

Each time we need to add new operation between client and server, we need to add
* two new messages to protobuf
* a member to `CommandResult`
* new rpc method
* two new macro calls `define_request!` and `result_convert!`
* new dumb server method implementation
* new client wrapper method in `FlushingBuckdClient`

Instead:
* Rust struct in `buck2_cli_proto`
  * which can use anything serde by the way, which makes things much easier
* two enum variants in the same file
* request type handling branch in `new_generic_command` function
* no changes in `.proto` files

That's it.

This diff converts `materialize` command to new universal request as an illustration.

Reviewed By: iguridi

Differential Revision: D49003048

fbshipit-source-id: aca935ce07cb1f31fe479d00570a829fe3285b5f