buck2: add a buck2-only linker wrapper rule that preserves exe / flags difference
Summary:
Doing this to help D48329688. For the Go rules, they want to be able to separate linker cmd and linker flags, but here we
have a bunch of linker flags that actually come via the linker cmd!

This fixes that by allowing the linker wrappers in fbcode to expose the separation between cmd and flags, and then the
toolchain implementation can place the flags where they belong.

Reviewed By: cjhopman

Differential Revision: D48899332

fbshipit-source-id: fdfd117996ae556d70b845b569ca1d6bb1ecb965