buck2: put panic back again
Summary:
Turns out is actually expected and very common for this to error out (100k times in 6 hours according to logview)

So put it back as it was before, ie. only panic for non Buck Events

Also add soft error for when it panics. It's a weird thing to do, so better be aware of how often it happens. I will followup next week

Reviewed By: cjhopman

Differential Revision: D49068063

fbshipit-source-id: 1a65d6996e2209af248fba3e0742bb435e842dda