rust-project should handle files with no owners gracefully
Summary:
Previously, we would error if none of the opened Rust files had associated targets. This would occur if the user only has files from the stdlib or generated files open.

 {F1085919603}

Instead, start rust-analyzer with a minimal config in this situation, so we provide some basic IDE functions and don't show any errors.

Reviewed By: JakobDegen

Differential Revision: D49077706

fbshipit-source-id: 1f399d12687f981666a7334970ec67af65163f4b