buck2: upload-logs: upload what we have after 5 min
Summary:
We upload in chunks of 8MiB. This means that until we get the first 8MiB of logs we won't upload nothing. Which is bad for hangs, since there could be logs to help with debugging. Yet we don't upload them when it's less than 8MiB of logs and the invocation never finishes gracefully.

This diff works around it by uploading what we have after 5 minutes without uploads

Reviewed By: stepancheg

Differential Revision: D49047169

fbshipit-source-id: 8205d401bdf62959971c821dec6abe6ccfb96562