remote_dep_file: Query remote dep file cache
Summary:
Part 2 from D48595643

This diff adds the change to process the remote dep file cache hit. (Remote dep file cache hit = an entry existed in the action cache under the remote dep file key). If there is a hit, the dep file entry must be inspected closer to check the action result can be used.

`check_remote_dep_file_entry` implements the logic for verifying if an action result is useable. The most important logic here is that we want to make sure all the inputs that are tracked by a dep file and shows up in the dep file (meaning it was used for the action) have not changed.

Reviewed By: scottcao

Differential Revision: D48605453

fbshipit-source-id: 60850cb8bfac94b9b797616f082cc2f173c64981