debug eval
Summary:
Evaluate `bzl` or `bxl` file, and that's it.

D48837424 adds unconditional typechecking for BXL.

Would be good to check that no BXL files break because of enabled typechecking — can be done with this command.

Additionally we may want this command in CI: on changes in buck2 trigger evaluation of every bxl file in the repository — it is cheap, and catches many errors.

JakobDegen also note now much simpler this diff is with new generic request.

This diff does not add command start/end, because it is not needed?

Reviewed By: JakobDegen

Differential Revision: D49076225

fbshipit-source-id: f68576c28bbe94ad8253e49a8d8206eb9a4e3b1e