Add a `link_group_deps` parameter
Summary:
D46022938 change inline `link_group_map` parsing to parse
the root as a `Label` instead of a `Dependency`.  While this
fixed the issues documented on the diff, it also broke the ability
to add link-group-only deps to the rule (the alternate, adding
these deps to `deps`, have the side-effect of them being linked
against the main binary).

This diff adds a new `link_group_deps` parameter to support
this use-case.  These deps are only used to augment the linkable
graph, but are not actual deps of the main executable.

Reviewed By: helfper

Differential Revision: D49335282

fbshipit-source-id: d7c2efd022bdb624cdf95ff800279ef4d436f00c