Make argument resolution not require the current directory
Summary: We have a context object containing the working directory, but we still use the current working directory. That makes the code test proof, and also means that if the two things ever diverge or we move argument parsing around it will break. Move to doing the resolution on the context object.

Reviewed By: stepancheg

Differential Revision: D49094225

fbshipit-source-id: 3f20e8dc861cf4140556a91aab65fe503bc36cab