Support cxx_precompiled_header() in link groups
Summary:
**Problem**

`cxx_precompiled_header()` can appear as a dependency of `(cxx|apple)_(binary|library)()`. The implementation trips up an internal check since there's a `MergedLinkInfo` provider but not a `LinkableGraph` provider.

**Solution**

Since `cxx_precompiled_header()` is a rule that affects compiles but not links, we can skip the node for linking purposes.

Reviewed By: chatura-atapattu

Differential Revision: D49317114

fbshipit-source-id: fc937e7c9b9c735d67a883d23313cb50277ebafe