buck2: aquery: make `deps(target)` work again
Summary:
Like it says in the title. This fails right now because we present
`deps(target)` with 2 nodes: the analysis node and the actions node, but we
only allow querying the analysis node later.

We don't actually need to query any analysis here (we never produce edges to
one), but an implementation detail of DFS means that we must (because we pass
keys even though we have their values).

So, handle this by looking up the keys we're asked for in the K/V set we had as
input. This way, we can't lookup arbitrary analysis nodes, but we can lookup
the one we passed in.

Reviewed By: wendy728

Differential Revision: D49408704

fbshipit-source-id: 337032bf1b9ddf3981393dcce9c9f806629cd1ac