Introduce cfg_constructor.bzl skeleton
Summary:
This diff adds the basic skeleton for cfg constructor functions used to evaluate modifiers. All it does right now is accept `legacy_platform` as an optional `PlatformInfo` and return a `PlatformInfo`.
Added setup for test case, which is setup via bxl. Basically, this acts like unit tests for testing the cfg constructor functions. However, we do run analysis of certain constraints in the repo to get their target labels and providers. In the future, this should be moved to isolated test once we have a better story on isolated test that actually tests prelude. The test case added tests that `PlatformInfo` is unchanged when there are no modifiers.
Not adding TODOs right now because this isn't used yet and most TODOs will be fixed by later diffs.

Differential Revision: D49358231

fbshipit-source-id: dba75d3a77ca7c03e36ba4ca16858598c7bb1a7e