Allow get_base_path() and package_name() from PACKAGE files
Summary:
Currently, it's only possible to call `get_base_path()` and `package_name()` from a buildfile. I want to be able to invoke these from a PACKAGE file to get the base path of PACKAGE file as well. For example, for `fbcode//foo/bar/PACKAGE` it would be `foo/bar`.
This is needed to record which PACKAGE a PACKAGE modifier came from when implementing PACKAGE modifier in D49483308.

Note this is just one idea I have in mind. I'm not opposed to hearing different ideas on how this can be done. One alternative idea I have is to add a `get_package_dir()` that's separate from `get_base_path()` and `package_name()` that is only invokable from PACKAGE files.

Reviewed By: stepancheg

Differential Revision: D49478298

fbshipit-source-id: 6decfa6de773070b124929ba8587575dfa41eaef