Promote configured_targets.rs to its own directory module ctargets
Summary:
Moving `ctargets` logic in `/buck2_server_commands` to its own directory.

This is set-up for adding `ctargets --keep-going` later, so let's encapsulate everything in one dir.

If this was the only change, I would leave it be. But `--streaming` might be added later and I also want to make changes to support JSON output. Potentially sizeable changes.

Reviewed By: bobyangyf

Differential Revision: D47523133

fbshipit-source-id: c1061afc9712297494b4ceb9ec884935173ab6b0