Add basic naming function
Summary: Because all configurations are generated by cfg constructor functions, they no longer have hand-specified names. This diff adds basic derivation of a name from a ConfigurationInfo. The process is pretty simple right now. We specify a list of key `constraint_setting` targets that should be named in the configuration. If the constraint setting is in the configuration, we add the name of the constraint value into the configuration name.

Differential Revision: D49359760

fbshipit-source-id: 8c4340558b197a7ba212f5b99dd192b504c4eccf