buck2_error: Derive macro
Summary: This diff provides the `#[buck2_error::Error]` attribute macro, which is a drop-in for `thiserror::Error`. In the short term, we use this to make some error types slightly more buckler compatible - the actual long term goal is to use this to support annotating errors with additional context information directly on the error definition.

Reviewed By: stepancheg

Differential Revision: D48672468

fbshipit-source-id: 14555bdefab602a56879cfd0df6fec0d7d4b516c