buck2_error: Introduce
Summary:
This diff introduces the initial implementation of the `buck2_error` crate. I'll start by briefly saying something about the overall implementation strategy for this and the related work:

The goal is to work vertically. This means doing end-to-end work on one improvement at a time, instead of trying to make things feature-complete from the beginning. As a result, the current `buckler::Error` lacks most of the features described in the RFC.

With respect to this diff:

The current version should mostly be fairly uncontroversial. We do provide the `AnyError` trait to support converting to `buck2_error::Error` from a broad range of error types; however, the docs state clearly that `anyhow` will remain the primary error type.

Note that although I do believe in what this diff says about `impl Context for Option<T>`, I did have to give up on that at least for the foreseeable future two diffs later

Reviewed By: stepancheg

Differential Revision: D48672466

fbshipit-source-id: c885cb5653949f4fde634929f276d89520885cdf