buck2_error: Make derive macros work in `buck2_error` crate
Summary: As written, the derive macros generate references to `::buck2_error`. This is going to generally be the right thing to have, but it doesn't work within the `buckler` crate itself. To work around this, we'll generate a second copy of the macros which use `crate` in place of `::buck2_error` - those will work within `buck2_error` and only be used there

Reviewed By: stepancheg

Differential Revision: D48672465

fbshipit-source-id: 62ca6b73542cc1d1a8b99111702baf375057834a