Add json type to cli args, supporting only primitives
Summary:
`cli_args.json()` only takes in a JSON with primitive types which map to serde_json::Values (so no targetexpr or subtargets). It's implemented such that the json must be an object, since I feel like a simpler implementation to expect a dict-like object, but open to changing to allow JSON arrays as well./

I probably will add some functionality to read in a file of JSON at some point.

Reviewed By: scottcao

Differential Revision: D49694594

fbshipit-source-id: cd4b0223a685f0e40d5628d6cf9a621a879ec01d